@@ -1,24 +1,25 @@
 module go.uber.org/zap/benchmarks
 
-go 1.13
+go 1.21
 
 replace go.uber.org/zap => ../
 
 require (
 	github.com/apex/log v1.9.0
-	github.com/benbjohnson/clock v1.2.0 // indirect
-	github.com/go-kit/kit v0.12.0
+	github.com/go-kit/log v0.2.1
+	github.com/rs/zerolog v1.30.0
+	github.com/sirupsen/logrus v1.9.3
+	go.uber.org/multierr v1.11.0
+	go.uber.org/zap v1.23.0
+	gopkg.in/inconshreveable/log15.v2 v2.16.0
+)
+
+require (
+	github.com/go-logfmt/logfmt v0.6.0 // indirect
 	github.com/go-stack/stack v1.8.1 // indirect
-	github.com/kr/pretty v0.3.0 // indirect
-	github.com/mattn/go-colorable v0.1.11 // indirect
-	github.com/rogpeppe/go-internal v1.8.0 // indirect
-	github.com/rs/zerolog v1.26.0
-	github.com/sirupsen/logrus v1.8.1
-	go.uber.org/goleak v1.1.12 // indirect
-	go.uber.org/multierr v1.7.0
-	go.uber.org/zap v1.19.1
-	golang.org/x/mod v0.5.1 // indirect
-	golang.org/x/sys v0.0.0-20211103235746-7861aae1554b // indirect
-	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
-	gopkg.in/inconshreveable/log15.v2 v2.0.0-20200109203555-b30bc20e4fd1
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.19 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	golang.org/x/sys v0.12.0 // indirect
+	golang.org/x/term v0.12.0 // indirect
 )
