@@ -1,9 +1,12 @@
 module go.uber.org/zap/tools
 
+require golang.org/x/vuln v1.0.1
+
 require (
-	golang.org/x/lint v0.0.0-20190930215403-16217165b5de
-	golang.org/x/tools v0.0.0-20191029190741-b9c20aec41a5 // indirect
-	honnef.co/go/tools v0.0.1-2019.2.3
+	golang.org/x/mod v0.12.0 // indirect
+	golang.org/x/sync v0.3.0 // indirect
+	golang.org/x/sys v0.11.0 // indirect
+	golang.org/x/tools v0.12.1-0.20230815132531-74c255bcf846 // indirect
 )
 
-go 1.13
+go 1.20
