@@ -21,13 +21,13 @@
 package zap
 
 import (
-	"io/ioutil"
 	"os"
+	"path/filepath"
+	"sync/atomic"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.uber.org/atomic"
 	"go.uber.org/zap/zapcore"
 )
 
@@ -42,27 +42,25 @@ func TestConfig(t *testing.T) {
 			desc:    "production",
 			cfg:     NewProductionConfig(),
 			expectN: 2 + 100 + 1, // 2 from initial logs, 100 initial sampled logs, 1 from off-by-one in sampler
-			expectRe: `{"level":"info","caller":"zap/config_test.go:\d+","msg":"info","k":"v","z":"zz"}` + "\n" +
-				`{"level":"warn","caller":"zap/config_test.go:\d+","msg":"warn","k":"v","z":"zz"}` + "\n",
+			expectRe: `{"level":"info","caller":"[a-z0-9_-]+/config_test.go:\d+","msg":"info","k":"v","z":"zz"}` + "\n" +
+				`{"level":"warn","caller":"[a-z0-9_-]+/config_test.go:\d+","msg":"warn","k":"v","z":"zz"}` + "\n",
 		},
 		{
 			desc:    "development",
 			cfg:     NewDevelopmentConfig(),
 			expectN: 3 + 200, // 3 initial logs, all 200 subsequent logs
-			expectRe: "DEBUG\tzap/config_test.go:" + `\d+` + "\tdebug\t" + `{"k": "v", "z": "zz"}` + "\n" +
-				"INFO\tzap/config_test.go:" + `\d+` + "\tinfo\t" + `{"k": "v", "z": "zz"}` + "\n" +
-				"WARN\tzap/config_test.go:" + `\d+` + "\twarn\t" + `{"k": "v", "z": "zz"}` + "\n" +
+			expectRe: "DEBUG\t[a-z0-9_-]+/config_test.go:" + `\d+` + "\tdebug\t" + `{"k": "v", "z": "zz"}` + "\n" +
+				"INFO\t[a-z0-9_-]+/config_test.go:" + `\d+` + "\tinfo\t" + `{"k": "v", "z": "zz"}` + "\n" +
+				"WARN\t[a-z0-9_-]+/config_test.go:" + `\d+` + "\twarn\t" + `{"k": "v", "z": "zz"}` + "\n" +
 				`go.uber.org/zap.TestConfig.\w+`,
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.desc, func(t *testing.T) {
-			temp, err := ioutil.TempFile("", "zap-prod-config-test")
-			require.NoError(t, err, "Failed to create temp file.")
-			defer os.Remove(temp.Name())
+			logOut := filepath.Join(t.TempDir(), "test.log")
 
-			tt.cfg.OutputPaths = []string{temp.Name()}
+			tt.cfg.OutputPaths = []string{logOut}
 			tt.cfg.EncoderConfig.TimeKey = "" // no timestamps in tests
 			tt.cfg.InitialFields = map[string]interface{}{"z": "zz", "k": "v"}
 
@@ -74,7 +72,7 @@ func TestConfig(t *testing.T) {
 			logger.Info("info")
 			logger.Warn("warn")
 
-			byteContents, err := ioutil.ReadAll(temp)
+			byteContents, err := os.ReadFile(logOut)
 			require.NoError(t, err, "Couldn't read log contents from temp file.")
 			logs := string(byteContents)
 			assert.Regexp(t, tt.expectRe, logs, "Unexpected log output.")
@@ -140,21 +138,21 @@ func TestConfigWithMissingAttributes(t *testing.T) {
 		t.Run(tt.desc, func(t *testing.T) {
 			cfg := tt.cfg
 			_, err := cfg.Build()
-			require.Error(t, err)
-			assert.Equal(t, tt.expectErr, err.Error())
+			assert.EqualError(t, err, tt.expectErr)
 		})
 	}
 }
 
 func makeSamplerCountingHook() (h func(zapcore.Entry, zapcore.SamplingDecision),
-	dropped, sampled *atomic.Int64) {
+	dropped, sampled *atomic.Int64,
+) {
 	dropped = new(atomic.Int64)
 	sampled = new(atomic.Int64)
 	h = func(_ zapcore.Entry, dec zapcore.SamplingDecision) {
 		if dec&zapcore.LogDropped > 0 {
-			dropped.Inc()
+			dropped.Add(1)
 		} else if dec&zapcore.LogSampled > 0 {
-			sampled.Inc()
+			sampled.Add(1)
 		}
 	}
 	return h, dropped, sampled
@@ -175,21 +173,13 @@ func TestConfigWithSamplingHook(t *testing.T) {
 		OutputPaths:      []string{"stderr"},
 		ErrorOutputPaths: []string{"stderr"},
 	}
-	expectRe := `{"level":"info","caller":"zap/config_test.go:\d+","msg":"info","k":"v","z":"zz"}` + "\n" +
-		`{"level":"warn","caller":"zap/config_test.go:\d+","msg":"warn","k":"v","z":"zz"}` + "\n"
+	expectRe := `{"level":"info","caller":"[a-z0-9_-]+/config_test.go:\d+","msg":"info","k":"v","z":"zz"}` + "\n" +
+		`{"level":"warn","caller":"[a-z0-9_-]+/config_test.go:\d+","msg":"warn","k":"v","z":"zz"}` + "\n"
 	expectDropped := 99  // 200 - 100 initial - 1 thereafter
 	expectSampled := 103 // 2 from initial + 100 + 1 thereafter
 
-	temp, err := ioutil.TempFile("", "zap-prod-config-test")
-	require.NoError(t, err, "Failed to create temp file.")
-	defer func() {
-		err := os.Remove(temp.Name())
-		if err != nil {
-			return
-		}
-	}()
-
-	cfg.OutputPaths = []string{temp.Name()}
+	logOut := filepath.Join(t.TempDir(), "test.log")
+	cfg.OutputPaths = []string{logOut}
 	cfg.EncoderConfig.TimeKey = "" // no timestamps in tests
 	cfg.InitialFields = map[string]interface{}{"z": "zz", "k": "v"}
 
@@ -200,7 +190,7 @@ func TestConfigWithSamplingHook(t *testing.T) {
 	logger.Info("info")
 	logger.Warn("warn")
 
-	byteContents, err := ioutil.ReadAll(temp)
+	byteContents, err := os.ReadFile(logOut)
 	require.NoError(t, err, "Couldn't read log contents from temp file.")
 	logs := string(byteContents)
 	assert.Regexp(t, expectRe, logs, "Unexpected log output.")
