@@ -22,7 +22,7 @@ package zap_test
 
 import (
 	"encoding/json"
-	"io/ioutil"
+	"io"
 	"log"
 	"os"
 	"time"
@@ -91,10 +91,7 @@ func Example_basicConfiguration() {
 	if err := json.Unmarshal(rawJSON, &cfg); err != nil {
 		panic(err)
 	}
-	logger, err := cfg.Build()
-	if err != nil {
-		panic(err)
-	}
+	logger := zap.Must(cfg.Build())
 	defer logger.Sync()
 
 	logger.Info("logger construction succeeded")
@@ -125,8 +122,8 @@ func Example_advancedConfiguration() {
 	// implement io.Writer, we can use zapcore.AddSync to add a no-op Sync
 	// method. If they're not safe for concurrent use, we can add a protecting
 	// mutex with zapcore.Lock.)
-	topicDebugging := zapcore.AddSync(ioutil.Discard)
-	topicErrors := zapcore.AddSync(ioutil.Discard)
+	topicDebugging := zapcore.AddSync(io.Discard)
+	topicErrors := zapcore.AddSync(io.Discard)
 
 	// High-priority output should also go to standard error, and low-priority
 	// output should also go to standard out.
@@ -182,7 +179,7 @@ func (a addr) MarshalLogObject(enc zapcore.ObjectEncoder) error {
 	return nil
 }
 
-func (r request) MarshalLogObject(enc zapcore.ObjectEncoder) error {
+func (r *request) MarshalLogObject(enc zapcore.ObjectEncoder) error {
 	enc.AddString("url", r.URL)
 	zap.Inline(r.Listen).AddTo(enc)
 	return enc.AddObject("remote", r.Remote)
@@ -278,10 +275,7 @@ func ExampleAtomicLevel_config() {
 	if err := json.Unmarshal(rawJSON, &cfg); err != nil {
 		panic(err)
 	}
-	logger, err := cfg.Build()
-	if err != nil {
-		panic(err)
-	}
+	logger := zap.Must(cfg.Build())
 	defer logger.Sync()
 
 	logger.Info("info logging enabled")
@@ -363,4 +357,57 @@ func ExampleWrapCore_wrap() {
 	// {"level":"info","msg":"single"}
 	// {"level":"info","msg":"doubled"}
 	// {"level":"info","msg":"doubled"}
+}
+
+func ExampleDict() {
+	logger := zap.NewExample()
+	defer logger.Sync()
+
+	logger.Info("login event",
+		zap.Dict("event",
+			zap.Int("id", 123),
+			zap.String("name", "jane"),
+			zap.String("status", "pending")))
+	// Output:
+	// {"level":"info","msg":"login event","event":{"id":123,"name":"jane","status":"pending"}}
+}
+
+func ExampleObjects() {
+	logger := zap.NewExample()
+	defer logger.Sync()
+
+	// Use the Objects field constructor when you have a list of objects,
+	// all of which implement zapcore.ObjectMarshaler.
+	logger.Debug("opening connections",
+		zap.Objects("addrs", []addr{
+			{IP: "123.45.67.89", Port: 4040},
+			{IP: "127.0.0.1", Port: 4041},
+			{IP: "192.168.0.1", Port: 4042},
+		}))
+	// Output:
+	// {"level":"debug","msg":"opening connections","addrs":[{"ip":"123.45.67.89","port":4040},{"ip":"127.0.0.1","port":4041},{"ip":"192.168.0.1","port":4042}]}
+}
+
+func ExampleObjectValues() {
+	logger := zap.NewExample()
+	defer logger.Sync()
+
+	// Use the ObjectValues field constructor when you have a list of
+	// objects that do not implement zapcore.ObjectMarshaler directly,
+	// but on their pointer receivers.
+	logger.Debug("starting tunnels",
+		zap.ObjectValues("addrs", []request{
+			{
+				URL:    "/foo",
+				Listen: addr{"127.0.0.1", 8080},
+				Remote: addr{"123.45.67.89", 4040},
+			},
+			{
+				URL:    "/bar",
+				Listen: addr{"127.0.0.1", 8080},
+				Remote: addr{"127.0.0.1", 31200},
+			},
+		}))
+	// Output:
+	// {"level":"debug","msg":"starting tunnels","addrs":[{"url":"/foo","ip":"127.0.0.1","port":8080,"remote":{"ip":"123.45.67.89","port":4040}},{"url":"/bar","ip":"127.0.0.1","port":8080,"remote":{"ip":"127.0.0.1","port":31200}}]}
 }
