@@ -18,12 +18,13 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+// readme generates Zap's README from a template.
 package main
 
 import (
 	"flag"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"log"
 	"os"
 	"os/exec"
@@ -34,18 +35,18 @@ import (
 	"time"
 )
 
-var (
-	libraryNameToMarkdownName = map[string]string{
-		"Zap":                   ":zap: zap",
-		"Zap.Sugar":             ":zap: zap (sugared)",
-		"stdlib.Println":        "standard library",
-		"sirupsen/logrus":       "logrus",
-		"go-kit/kit/log":        "go-kit",
-		"inconshreveable/log15": "log15",
-		"apex/log":              "apex/log",
-		"rs/zerolog":            "zerolog",
-	}
-)
+var libraryNameToMarkdownName = map[string]string{
+	"Zap":                   ":zap: zap",
+	"Zap.Sugar":             ":zap: zap (sugared)",
+	"stdlib.Println":        "standard library",
+	"sirupsen/logrus":       "logrus",
+	"go-kit/kit/log":        "go-kit",
+	"inconshreveable/log15": "log15",
+	"apex/log":              "apex/log",
+	"rs/zerolog":            "zerolog",
+	"slog":                  "slog",
+	"slog.LogAttrs":         "slog (LogAttrs)",
+}
 
 func main() {
 	flag.Parse()
@@ -59,7 +60,7 @@ func do() error {
 	if err != nil {
 		return err
 	}
-	data, err := ioutil.ReadAll(os.Stdin)
+	data, err := io.ReadAll(os.Stdin)
 	if err != nil {
 		return err
 	}
@@ -140,7 +141,7 @@ func getBenchmarkRow(
 	if len(split) < 5 {
 		return nil, fmt.Errorf("unknown benchmark line: %s", line)
 	}
-	duration, err := time.ParseDuration(strings.Replace(strings.TrimSuffix(strings.TrimSpace(split[2]), "/op"), " ", "", -1))
+	duration, err := time.ParseDuration(strings.ReplaceAll(strings.TrimSuffix(strings.TrimSpace(split[2]), "/op"), " ", ""))
 	if err != nil {
 		return nil, err
 	}
