@@ -21,7 +21,6 @@
 package zapcore
 
 import (
-	"io/ioutil"
 	"os"
 	"testing"
 
@@ -31,22 +30,25 @@ import (
 
 func BenchmarkBufferedWriteSyncer(b *testing.B) {
 	b.Run("write file with buffer", func(b *testing.B) {
-		file, err := ioutil.TempFile("", "log")
+		file, err := os.CreateTemp(b.TempDir(), "test.log")
 		require.NoError(b, err)
 
 		defer func() {
 			assert.NoError(b, file.Close())
-			assert.NoError(b, os.Remove(file.Name()))
 		}()
 
 		w := &BufferedWriteSyncer{
 			WS: AddSync(file),
 		}
-		defer w.Stop()
+		defer func() {
+			assert.NoError(b, w.Stop(), "failed to stop buffered write syncer")
+		}()
 		b.ResetTimer()
 		b.RunParallel(func(pb *testing.PB) {
 			for pb.Next() {
-				w.Write([]byte("foobarbazbabble"))
+				if _, err := w.Write([]byte("foobarbazbabble")); err != nil {
+					b.Fatal(err)
+				}
 			}
 		})
 	})
