@@ -22,12 +22,12 @@ package zapcore_test
 
 import (
 	"errors"
-	"io/ioutil"
 	"os"
 	"testing"
 	"time"
 
 	"go.uber.org/zap/internal/ztest"
+	//revive:disable:dot-imports
 	. "go.uber.org/zap/zapcore"
 
 	"github.com/stretchr/testify/assert"
@@ -68,9 +68,8 @@ func TestNopCore(t *testing.T) {
 }
 
 func TestIOCore(t *testing.T) {
-	temp, err := ioutil.TempFile("", "zapcore-test-iocore")
-	require.NoError(t, err, "Failed to create temp file.")
-	defer os.Remove(temp.Name())
+	temp, err := os.CreateTemp(t.TempDir(), "test.log")
+	require.NoError(t, err)
 
 	// Drop timestamps for simpler assertions (timestamp encoding is tested
 	// elsewhere).
@@ -84,6 +83,10 @@ func TestIOCore(t *testing.T) {
 	).With([]Field{makeInt64Field("k", 1)})
 	defer assert.NoError(t, core.Sync(), "Expected Syncing a temp file to succeed.")
 
+	t.Run("LevelOf", func(t *testing.T) {
+		assert.Equal(t, InfoLevel, LevelOf(core), "Incorrect Core Level")
+	})
+
 	if ce := core.Check(Entry{Level: DebugLevel, Message: "debug"}, nil); ce != nil {
 		ce.Write(makeInt64Field("k", 2))
 	}
@@ -94,7 +97,7 @@ func TestIOCore(t *testing.T) {
 		ce.Write(makeInt64Field("k", 4))
 	}
 
-	logged, err := ioutil.ReadFile(temp.Name())
+	logged, err := os.ReadFile(temp.Name())
 	require.NoError(t, err, "Failed to read from temp file.")
 	require.Equal(
 		t,
@@ -146,7 +149,7 @@ func TestIOCoreSyncsOutput(t *testing.T) {
 			DebugLevel,
 		)
 
-		core.Write(tt.entry, nil)
+		assert.NoError(t, core.Write(tt.entry, nil), "Unexpected error writing entry.")
 		assert.Equal(t, tt.shouldSync, sink.Called(), "Incorrect Sync behavior.")
 	}
 }
