@@ -21,11 +21,13 @@
 package benchmarks
 
 import (
-	"io/ioutil"
+	"context"
 	"log"
+	"log/slog"
 	"testing"
 
 	"go.uber.org/zap"
+	"go.uber.org/zap/internal/ztest"
 )
 
 func BenchmarkDisabledWithoutFields(b *testing.B) {
@@ -92,6 +94,24 @@ func BenchmarkDisabledWithoutFields(b *testing.B) {
 		b.RunParallel(func(pb *testing.PB) {
 			for pb.Next() {
 				logger.Info().Msg(getMessage(0))
+			}
+		})
+	})
+	b.Run("slog", func(b *testing.B) {
+		logger := newDisabledSlog()
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.Info(getMessage(0))
+			}
+		})
+	})
+	b.Run("slog.LogAttrs", func(b *testing.B) {
+		logger := newDisabledSlog()
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.LogAttrs(context.Background(), slog.LevelInfo, getMessage(0))
 			}
 		})
 	})
@@ -164,6 +184,24 @@ func BenchmarkDisabledAccumulatedContext(b *testing.B) {
 			}
 		})
 	})
+	b.Run("slog", func(b *testing.B) {
+		logger := newDisabledSlog(fakeSlogFields()...)
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.Info(getMessage(0))
+			}
+		})
+	})
+	b.Run("slog.LogAttrs", func(b *testing.B) {
+		logger := newDisabledSlog(fakeSlogFields()...)
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.LogAttrs(context.Background(), slog.LevelInfo, getMessage(0))
+			}
+		})
+	})
 }
 
 func BenchmarkDisabledAddingFields(b *testing.B) {
@@ -221,6 +259,24 @@ func BenchmarkDisabledAddingFields(b *testing.B) {
 		b.RunParallel(func(pb *testing.PB) {
 			for pb.Next() {
 				fakeZerologFields(logger.Info()).Msg(getMessage(0))
+			}
+		})
+	})
+	b.Run("slog", func(b *testing.B) {
+		logger := newDisabledSlog()
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.Info(getMessage(0), fakeSlogArgs()...)
+			}
+		})
+	})
+	b.Run("slog.LogAttrs", func(b *testing.B) {
+		logger := newDisabledSlog()
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.LogAttrs(context.Background(), slog.LevelInfo, getMessage(0), fakeSlogFields()...)
 			}
 		})
 	})
@@ -293,7 +349,9 @@ func BenchmarkWithoutFields(b *testing.B) {
 		b.ResetTimer()
 		b.RunParallel(func(pb *testing.PB) {
 			for pb.Next() {
-				logger.Log(getMessage(0), getMessage(1))
+				if err := logger.Log(getMessage(0), getMessage(1)); err != nil {
+					b.Fatal(err)
+				}
 			}
 		})
 	})
@@ -316,7 +374,7 @@ func BenchmarkWithoutFields(b *testing.B) {
 		})
 	})
 	b.Run("stdlib.Println", func(b *testing.B) {
-		logger := log.New(ioutil.Discard, "", log.LstdFlags)
+		logger := log.New(&ztest.Discarder{}, "", log.LstdFlags)
 		b.ResetTimer()
 		b.RunParallel(func(pb *testing.PB) {
 			for pb.Next() {
@@ -325,7 +383,7 @@ func BenchmarkWithoutFields(b *testing.B) {
 		})
 	})
 	b.Run("stdlib.Printf", func(b *testing.B) {
-		logger := log.New(ioutil.Discard, "", log.LstdFlags)
+		logger := log.New(&ztest.Discarder{}, "", log.LstdFlags)
 		b.ResetTimer()
 		b.RunParallel(func(pb *testing.PB) {
 			for pb.Next() {
@@ -359,6 +417,24 @@ func BenchmarkWithoutFields(b *testing.B) {
 				if e := logger.Info(); e.Enabled() {
 					e.Msg(getMessage(0))
 				}
+			}
+		})
+	})
+	b.Run("slog", func(b *testing.B) {
+		logger := newSlog()
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.Info(getMessage(0))
+			}
+		})
+	})
+	b.Run("slog.LogAttrs", func(b *testing.B) {
+		logger := newSlog()
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.LogAttrs(context.Background(), slog.LevelInfo, getMessage(0))
 			}
 		})
 	})
@@ -431,7 +507,9 @@ func BenchmarkAccumulatedContext(b *testing.B) {
 		b.ResetTimer()
 		b.RunParallel(func(pb *testing.PB) {
 			for pb.Next() {
-				logger.Log(getMessage(0), getMessage(1))
+				if err := logger.Log(getMessage(0), getMessage(1)); err != nil {
+					b.Fatal(err)
+				}
 			}
 		})
 	})
@@ -479,6 +557,24 @@ func BenchmarkAccumulatedContext(b *testing.B) {
 		b.RunParallel(func(pb *testing.PB) {
 			for pb.Next() {
 				logger.Info().Msgf("%v %v %v %s %v %v %v %v %v %s\n", fakeFmtArgs()...)
+			}
+		})
+	})
+	b.Run("slog", func(b *testing.B) {
+		logger := newSlog(fakeSlogFields()...)
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.Info(getMessage(0))
+			}
+		})
+	})
+	b.Run("slog.LogAttrs", func(b *testing.B) {
+		logger := newSlog(fakeSlogFields()...)
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.LogAttrs(context.Background(), slog.LevelInfo, getMessage(0))
 			}
 		})
 	})
@@ -542,7 +638,9 @@ func BenchmarkAddingFields(b *testing.B) {
 		b.ResetTimer()
 		b.RunParallel(func(pb *testing.PB) {
 			for pb.Next() {
-				logger.Log(fakeSugarFields()...)
+				if err := logger.Log(fakeSugarFields()...); err != nil {
+					b.Fatal(err)
+				}
 			}
 		})
 	})
@@ -581,6 +679,24 @@ func BenchmarkAddingFields(b *testing.B) {
 				if e := logger.Info(); e.Enabled() {
 					fakeZerologFields(e).Msg(getMessage(0))
 				}
+			}
+		})
+	})
+	b.Run("slog", func(b *testing.B) {
+		logger := newSlog()
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.Info(getMessage(0), fakeSlogArgs()...)
+			}
+		})
+	})
+	b.Run("slog.LogAttrs", func(b *testing.B) {
+		logger := newSlog()
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			for pb.Next() {
+				logger.LogAttrs(context.Background(), slog.LevelInfo, getMessage(0), fakeSlogFields()...)
 			}
 		})
 	})
