@@ -22,11 +22,11 @@ package zap
 
 import (
 	"errors"
+	"fmt"
 	"testing"
 
 	"go.uber.org/zap/zapcore"
 
-	richErrors "github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -79,7 +79,7 @@ func TestErrorArrayConstructor(t *testing.T) {
 }
 
 func TestErrorsArraysHandleRichErrors(t *testing.T) {
-	errs := []error{richErrors.New("egad")}
+	errs := []error{fmt.Errorf("egad")}
 
 	enc := zapcore.NewMapObjectEncoder()
 	Errors("k", errs).AddTo(enc)
@@ -94,6 +94,40 @@ func TestErrorsArraysHandleRichErrors(t *testing.T) {
 	errMap, ok := serialized.(map[string]interface{})
 	require.True(t, ok, "Expected serialized error to be a map, got %T.", serialized)
 	assert.Equal(t, "egad", errMap["error"], "Unexpected standard error string.")
-	assert.Contains(t, errMap["errorVerbose"], "egad", "Verbose error string should be a superset of standard error.")
-	assert.Contains(t, errMap["errorVerbose"], "TestErrorsArraysHandleRichErrors", "Verbose error string should contain a stacktrace.")
+}
+
+func TestErrArrayBrokenEncoder(t *testing.T) {
+	t.Parallel()
+
+	failWith := errors.New("great sadness")
+	err := (brokenArrayObjectEncoder{
+		Err:           failWith,
+		ObjectEncoder: zapcore.NewMapObjectEncoder(),
+	}).AddArray("errors", errArray{
+		errors.New("foo"),
+		errors.New("bar"),
+	})
+	require.Error(t, err, "Expected error from broken encoder.")
+	assert.ErrorIs(t, err, failWith, "Unexpected error.")
+}
+
+// brokenArrayObjectEncoder is an ObjectEncoder
+// that builds a broken ArrayEncoder.
+type brokenArrayObjectEncoder struct {
+	zapcore.ObjectEncoder
+	zapcore.ArrayEncoder
+
+	Err error // error to return
+}
+
+func (enc brokenArrayObjectEncoder) AddArray(key string, marshaler zapcore.ArrayMarshaler) error {
+	return enc.ObjectEncoder.AddArray(key,
+		zapcore.ArrayMarshalerFunc(func(ae zapcore.ArrayEncoder) error {
+			enc.ArrayEncoder = ae
+			return marshaler.MarshalLogArray(enc)
+		}))
+}
+
+func (enc brokenArrayObjectEncoder) AppendObject(zapcore.ObjectMarshaler) error {
+	return enc.Err
 }
