@@ -22,7 +22,10 @@ package zap
 
 import (
 	"errors"
+	"fmt"
+	"strconv"
 	"sync"
+	"sync/atomic"
 	"testing"
 
 	"go.uber.org/zap/internal/exit"
@@ -32,13 +35,12 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.uber.org/atomic"
 )
 
 func makeCountingHook() (func(zapcore.Entry) error, *atomic.Int64) {
 	count := &atomic.Int64{}
 	h := func(zapcore.Entry) error {
-		count.Inc()
+		count.Add(1)
 		return nil
 	}
 	return h, count
@@ -83,6 +85,33 @@ func TestLoggerAtomicLevel(t *testing.T) {
 	})
 }
 
+func TestLoggerLevel(t *testing.T) {
+	levels := []zapcore.Level{
+		DebugLevel,
+		InfoLevel,
+		WarnLevel,
+		ErrorLevel,
+		DPanicLevel,
+		PanicLevel,
+		FatalLevel,
+	}
+
+	for _, lvl := range levels {
+		lvl := lvl
+		t.Run(lvl.String(), func(t *testing.T) {
+			t.Parallel()
+
+			core, _ := observer.New(lvl)
+			log := New(core)
+			assert.Equal(t, lvl, log.Level())
+		})
+	}
+
+	t.Run("Nop", func(t *testing.T) {
+		assert.Equal(t, zapcore.InvalidLevel, NewNop().Level())
+	})
+}
+
 func TestLoggerInitialFields(t *testing.T) {
 	fieldOpts := opts(Fields(Int("foo", 42), String("bar", "baz")))
 	withLogger(t, DebugLevel, fieldOpts, func(logger *Logger, logs *observer.ObservedLogs) {
@@ -97,20 +126,250 @@ func TestLoggerInitialFields(t *testing.T) {
 }
 
 func TestLoggerWith(t *testing.T) {
-	fieldOpts := opts(Fields(Int("foo", 42)))
-	withLogger(t, DebugLevel, fieldOpts, func(logger *Logger, logs *observer.ObservedLogs) {
-		// Child loggers should have copy-on-write semantics, so two children
-		// shouldn't stomp on each other's fields or affect the parent's fields.
-		logger.With(String("one", "two")).Info("")
-		logger.With(String("three", "four")).Info("")
-		logger.Info("")
+	tests := []struct {
+		name          string
+		initialFields []Field
+		withMethod    func(*Logger, ...Field) *Logger
+	}{
+		{
+			"regular non lazy logger",
+			[]Field{Int("foo", 42)},
+			(*Logger).With,
+		},
+		{
+			"regular non lazy logger no initial fields",
+			[]Field{},
+			(*Logger).With,
+		},
+		{
+			"lazy with logger",
+			[]Field{Int("foo", 42)},
+			(*Logger).WithLazy,
+		},
+		{
+			"lazy with logger no initial fields",
+			[]Field{},
+			(*Logger).WithLazy,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			withLogger(t, DebugLevel, opts(Fields(tt.initialFields...)), func(logger *Logger, logs *observer.ObservedLogs) {
+				// Child loggers should have copy-on-write semantics, so two children
+				// shouldn't stomp on each other's fields or affect the parent's fields.
+				tt.withMethod(logger).Info("")
+				tt.withMethod(logger, String("one", "two")).Info("")
+				tt.withMethod(logger, String("three", "four")).Info("")
+				tt.withMethod(logger, String("five", "six")).With(String("seven", "eight")).Info("")
+				logger.Info("")
 
-		assert.Equal(t, []observer.LoggedEntry{
-			{Context: []Field{Int("foo", 42), String("one", "two")}},
-			{Context: []Field{Int("foo", 42), String("three", "four")}},
-			{Context: []Field{Int("foo", 42)}},
-		}, logs.AllUntimed(), "Unexpected cross-talk between child loggers.")
-	})
+				assert.Equal(t, []observer.LoggedEntry{
+					{Context: tt.initialFields},
+					{Context: append(tt.initialFields, String("one", "two"))},
+					{Context: append(tt.initialFields, String("three", "four"))},
+					{Context: append(tt.initialFields, String("five", "six"), String("seven", "eight"))},
+					{Context: tt.initialFields},
+				}, logs.AllUntimed(), "Unexpected cross-talk between child loggers.")
+			})
+		})
+	}
+}
+
+func TestLoggerWithCaptures(t *testing.T) {
+	type withF func(*Logger, ...Field) *Logger
+	tests := []struct {
+		name        string
+		withMethods []withF
+		wantJSON    []string
+	}{
+		{
+			name:        "regular with captures arguments at time of With",
+			withMethods: []withF{(*Logger).With},
+			wantJSON: []string{
+				`{
+					"m": "hello 0",
+					"a0": [0],
+					"b0": [1]
+				}`,
+				`{
+					"m": "world 0",
+					"a0": [0],
+					"c0": [2]
+				}`,
+			},
+		},
+		{
+			name:        "lazy with captures arguments at time of With or Logging",
+			withMethods: []withF{(*Logger).WithLazy},
+			wantJSON: []string{
+				`{
+					"m": "hello 0",
+					"a0": [1],
+					"b0": [1]
+				}`,
+				`{
+					"m": "world 0",
+					"a0": [1],
+					"c0": [2]
+				}`,
+			},
+		},
+		{
+			name:        "2x With captures arguments at time of each With",
+			withMethods: []withF{(*Logger).With, (*Logger).With},
+			wantJSON: []string{
+				`{
+					"m": "hello 0",
+					"a0": [0],
+					"b0": [1]
+				}`,
+				`{
+					"m": "world 0",
+					"a0": [0],
+					"c0": [2]
+				}`,
+				`{
+					"m": "hello 1",
+					"a0": [0],
+					"c0": [2],
+					"a1": [10],
+					"b1": [11]
+				}`,
+				`{
+					"m": "world 1",
+					"a0": [0],
+					"c0": [2],
+					"a1": [10],
+					"c1": [12]
+				}`,
+			},
+		},
+		{
+			name:        "2x WithLazy. Captures arguments only at logging time.",
+			withMethods: []withF{(*Logger).WithLazy, (*Logger).WithLazy},
+			wantJSON: []string{
+				`{
+					"m": "hello 0",
+					"a0": [1],
+					"b0": [1]
+				}`,
+				`{
+					"m": "world 0",
+					"a0": [1],
+					"c0": [2]
+				}`,
+				`{
+					"m": "hello 1",
+					"a0": [1],
+					"c0": [2],
+					"a1": [11],
+					"b1": [11]
+				}`,
+				`{
+					"m": "world 1",
+					"a0": [1],
+					"c0": [2],
+					"a1": [11],
+					"c1": [12]
+				}`,
+			},
+		},
+		{
+			name:        "WithLazy then With",
+			withMethods: []withF{(*Logger).WithLazy, (*Logger).With},
+			wantJSON: []string{
+				`{
+					"m": "hello 0",
+					"a0": [1],
+					"b0": [1]
+				}`,
+				`{
+					"m": "world 0",
+					"a0": [1],
+					"c0": [2]
+				}`,
+				`{
+					"m": "hello 1",
+					"a0": [1],
+					"c0": [2],
+					"a1": [10],
+					"b1": [11]
+				}`,
+				`{
+					"m": "world 1",
+					"a0": [1],
+					"c0": [2],
+					"a1": [10],
+					"c1": [12]
+				}`,
+			},
+		},
+		{
+			name:        "With then WithLazy",
+			withMethods: []withF{(*Logger).With, (*Logger).WithLazy},
+			wantJSON: []string{
+				`{
+					"m": "hello 0",
+					"a0": [0],
+					"b0": [1]
+				}`,
+				`{
+					"m": "world 0",
+					"a0": [0],
+					"c0": [2]
+				}`,
+				`{
+					"m": "hello 1",
+					"a0": [0],
+					"c0": [2],
+					"a1": [11],
+					"b1": [11]
+				}`,
+				`{
+					"m": "world 1",
+					"a0": [0],
+					"c0": [2],
+					"a1": [11],
+					"c1": [12]
+				}`,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			enc := zapcore.NewJSONEncoder(zapcore.EncoderConfig{
+				MessageKey: "m",
+			})
+
+			var bs ztest.Buffer
+			logger := New(zapcore.NewCore(enc, &bs, DebugLevel))
+
+			for i, withMethod := range tt.withMethods {
+
+				iStr := strconv.Itoa(i)
+				x := 10 * i
+				arr := zapcore.ArrayMarshalerFunc(func(enc zapcore.ArrayEncoder) error {
+					enc.AppendInt(x)
+					return nil
+				})
+
+				// Demonstrate the arguments are captured when With() and Info() are invoked.
+				logger = withMethod(logger, Array("a"+iStr, arr))
+				x++
+				logger.Info(fmt.Sprintf("hello %d", i), Array("b"+iStr, arr))
+				x++
+				logger = withMethod(logger, Array("c"+iStr, arr))
+				logger.Info(fmt.Sprintf("world %d", i))
+			}
+
+			if lines := bs.Lines(); assert.Len(t, lines, len(tt.wantJSON)) {
+				for i, want := range tt.wantJSON {
+					assert.JSONEq(t, want, lines[i], "Unexpected output from the %d'th log.", i)
+				}
+			}
+		})
+	}
 }
 
 func TestLoggerLogPanic(t *testing.T) {
@@ -119,7 +378,8 @@ func TestLoggerLogPanic(t *testing.T) {
 		should   bool
 		expected string
 	}{
-		{func(logger *Logger) { logger.Check(PanicLevel, "bar").Write() }, true, "bar"},
+		{func(logger *Logger) { logger.Check(PanicLevel, "foo").Write() }, true, "foo"},
+		{func(logger *Logger) { logger.Log(PanicLevel, "bar") }, true, "bar"},
 		{func(logger *Logger) { logger.Panic("baz") }, true, "baz"},
 	} {
 		withLogger(t, DebugLevel, nil, func(logger *Logger, logs *observer.ObservedLogs) {
@@ -147,7 +407,8 @@ func TestLoggerLogFatal(t *testing.T) {
 		do       func(*Logger)
 		expected string
 	}{
-		{func(logger *Logger) { logger.Check(FatalLevel, "bar").Write() }, "bar"},
+		{func(logger *Logger) { logger.Check(FatalLevel, "foo").Write() }, "foo"},
+		{func(logger *Logger) { logger.Log(FatalLevel, "bar") }, "bar"},
 		{func(logger *Logger) { logger.Fatal("baz") }, "baz"},
 	} {
 		withLogger(t, DebugLevel, nil, func(logger *Logger, logs *observer.ObservedLogs) {
@@ -194,12 +455,36 @@ func TestLoggerLeveledMethods(t *testing.T) {
 	})
 }
 
+func TestLoggerLogLevels(t *testing.T) {
+	withLogger(t, DebugLevel, nil, func(logger *Logger, logs *observer.ObservedLogs) {
+		levels := []zapcore.Level{
+			DebugLevel,
+			InfoLevel,
+			WarnLevel,
+			ErrorLevel,
+			DPanicLevel,
+		}
+		for i, level := range levels {
+			logger.Log(level, "")
+			output := logs.AllUntimed()
+			assert.Equal(t, i+1, len(output), "Unexpected number of logs.")
+			assert.Equal(t, 0, len(output[i].Context), "Unexpected context on first log.")
+			assert.Equal(
+				t,
+				zapcore.Entry{Level: level},
+				output[i].Entry,
+				"Unexpected output from %s-level logger method.", level)
+		}
+	})
+}
+
 func TestLoggerAlwaysPanics(t *testing.T) {
 	// Users can disable writing out panic-level logs, but calls to logger.Panic()
 	// should still call panic().
 	withLogger(t, FatalLevel, nil, func(logger *Logger, logs *observer.ObservedLogs) {
 		msg := "Even if output is disabled, logger.Panic should always panic."
 		assert.Panics(t, func() { logger.Panic("foo") }, msg)
+		assert.Panics(t, func() { logger.Log(PanicLevel, "foo") }, msg)
 		assert.Panics(t, func() {
 			if ce := logger.Check(PanicLevel, "foo"); ce != nil {
 				ce.Write()
@@ -214,6 +499,9 @@ func TestLoggerAlwaysFatals(t *testing.T) {
 	// should still terminate the process.
 	withLogger(t, FatalLevel+1, nil, func(logger *Logger, logs *observer.ObservedLogs) {
 		stub := exit.WithStub(func() { logger.Fatal("") })
+		assert.True(t, stub.Exited, "Expected calls to logger.Fatal to terminate process.")
+
+		stub = exit.WithStub(func() { logger.Log(FatalLevel, "") })
 		assert.True(t, stub.Exited, "Expected calls to logger.Fatal to terminate process.")
 
 		stub = exit.WithStub(func() {
@@ -284,7 +572,8 @@ func TestLoggerNames(t *testing.T) {
 			}
 			log.Info("")
 			require.Equal(t, 1, logs.Len(), "Expected only one log entry to be written.")
-			assert.Equal(t, tt.expected, logs.AllUntimed()[0].Entry.LoggerName, "Unexpected logger name.")
+			assert.Equal(t, tt.expected, logs.AllUntimed()[0].LoggerName, "Unexpected logger name from entry.")
+			assert.Equal(t, tt.expected, log.Name(), "Unexpected logger name.")
 		})
 		withSugar(t, DebugLevel, nil, func(log *SugaredLogger, logs *observer.ObservedLogs) {
 			for _, n := range tt.names {
@@ -292,7 +581,8 @@ func TestLoggerNames(t *testing.T) {
 			}
 			log.Infow("")
 			require.Equal(t, 1, logs.Len(), "Expected only one log entry to be written.")
-			assert.Equal(t, tt.expected, logs.AllUntimed()[0].Entry.LoggerName, "Unexpected logger name.")
+			assert.Equal(t, tt.expected, logs.AllUntimed()[0].LoggerName, "Unexpected logger name from entry.")
+			assert.Equal(t, tt.expected, log.base.Name(), "Unexpected logger name.")
 		})
 	}
 }
@@ -345,8 +635,8 @@ func TestLoggerAddCaller(t *testing.T) {
 		{opts(AddCaller(), WithCaller(false)), `^undefined$`},
 		{opts(WithCaller(true)), `.+/logger_test.go:[\d]+$`},
 		{opts(WithCaller(true), WithCaller(false)), `^undefined$`},
-		{opts(AddCaller(), AddCallerSkip(1), AddCallerSkip(-1)), `.+/zap/logger_test.go:[\d]+$`},
-		{opts(AddCaller(), AddCallerSkip(1)), `.+/zap/common_test.go:[\d]+$`},
+		{opts(AddCaller(), AddCallerSkip(1), AddCallerSkip(-1)), `.+/logger_test.go:[\d]+$`},
+		{opts(AddCaller(), AddCallerSkip(1)), `.+/common_test.go:[\d]+$`},
 		{opts(AddCaller(), AddCallerSkip(1), AddCallerSkip(3)), `.+/src/runtime/.*:[\d]+$`},
 	}
 	for _, tt := range tests {
@@ -359,7 +649,7 @@ func TestLoggerAddCaller(t *testing.T) {
 			assert.Regexp(
 				t,
 				tt.pat,
-				output[0].Entry.Caller,
+				output[0].Caller,
 				"Expected to find package name and file name in output.",
 			)
 		})
@@ -432,14 +722,14 @@ func TestLoggerAddCallerFunction(t *testing.T) {
 				assert.Regexp(
 					t,
 					tt.loggerFunction,
-					entry.Entry.Caller.Function,
+					entry.Caller.Function,
 					"Expected to find function name in output.",
 				)
 			}
 			assert.Regexp(
 				t,
 				tt.sugaredFunction,
-				entries[1].Entry.Caller.Function,
+				entries[1].Caller.Function,
 				"Expected to find function name in output.",
 			)
 		})
@@ -458,12 +748,12 @@ func TestLoggerAddCallerFail(t *testing.T) {
 		)
 		assert.Equal(
 			t,
-			logs.AllUntimed()[0].Entry.Message,
+			logs.AllUntimed()[0].Message,
 			"Failure.",
 			"Expected original message to survive failures in runtime.Caller.")
 		assert.Equal(
 			t,
-			logs.AllUntimed()[0].Entry.Caller.Function,
+			logs.AllUntimed()[0].Caller.Function,
 			"",
 			"Expected function name to be empty string.")
 	})
@@ -489,7 +779,7 @@ func TestLoggerIncreaseLevel(t *testing.T) {
 		require.Equal(t, 2, logs.Len(), "expected only warn + error logs due to IncreaseLevel.")
 		assert.Equal(
 			t,
-			logs.AllUntimed()[0].Entry.Message,
+			logs.AllUntimed()[0].Message,
 			"logger.Warn",
 			"Expected first logged message to be warn level message",
 		)
@@ -533,6 +823,141 @@ func TestLoggerConcurrent(t *testing.T) {
 			)
 		}
 	})
+}
+
+func TestLoggerFatalOnNoop(t *testing.T) {
+	exitStub := exit.Stub()
+	defer exitStub.Unstub()
+	core, _ := observer.New(InfoLevel)
+
+	// We don't allow a no-op fatal hook.
+	New(core, WithFatalHook(zapcore.WriteThenNoop)).Fatal("great sadness")
+	assert.True(t, exitStub.Exited, "must exit for WriteThenNoop")
+	assert.Equal(t, 1, exitStub.Code, "must exit with status 1 for WriteThenNoop")
+}
+
+func TestLoggerCustomOnPanic(t *testing.T) {
+	tests := []struct {
+		msg          string
+		level        zapcore.Level
+		opts         []Option
+		finished     bool
+		want         []observer.LoggedEntry
+		recoverValue any
+	}{
+		{
+			msg:      "panic with nil hook",
+			level:    PanicLevel,
+			opts:     opts(WithPanicHook(nil)),
+			finished: false,
+			want: []observer.LoggedEntry{
+				{
+					Entry:   zapcore.Entry{Level: PanicLevel, Message: "foobar"},
+					Context: []Field{},
+				},
+			},
+			recoverValue: "foobar",
+		},
+		{
+			msg:      "panic with noop hook",
+			level:    PanicLevel,
+			opts:     opts(WithPanicHook(zapcore.WriteThenNoop)),
+			finished: false,
+			want: []observer.LoggedEntry{
+				{
+					Entry:   zapcore.Entry{Level: PanicLevel, Message: "foobar"},
+					Context: []Field{},
+				},
+			},
+			recoverValue: "foobar",
+		},
+		{
+			msg:      "no panic with goexit hook",
+			level:    PanicLevel,
+			opts:     opts(WithPanicHook(zapcore.WriteThenGoexit)),
+			finished: false,
+			want: []observer.LoggedEntry{
+				{
+					Entry:   zapcore.Entry{Level: PanicLevel, Message: "foobar"},
+					Context: []Field{},
+				},
+			},
+			recoverValue: nil,
+		},
+		{
+			msg:      "dpanic no panic in development mode with goexit hook",
+			level:    DPanicLevel,
+			opts:     opts(WithPanicHook(zapcore.WriteThenGoexit), Development()),
+			finished: false,
+			want: []observer.LoggedEntry{
+				{
+					Entry:   zapcore.Entry{Level: DPanicLevel, Message: "foobar"},
+					Context: []Field{},
+				},
+			},
+			recoverValue: nil,
+		},
+		{
+			msg:      "dpanic panic in development mode with noop hook",
+			level:    DPanicLevel,
+			opts:     opts(WithPanicHook(zapcore.WriteThenNoop), Development()),
+			finished: false,
+			want: []observer.LoggedEntry{
+				{
+					Entry:   zapcore.Entry{Level: DPanicLevel, Message: "foobar"},
+					Context: []Field{},
+				},
+			},
+			recoverValue: "foobar",
+		},
+		{
+			msg:      "dpanic no exit in production mode with goexit hook",
+			level:    DPanicLevel,
+			opts:     opts(WithPanicHook(zapcore.WriteThenPanic)),
+			finished: true,
+			want: []observer.LoggedEntry{
+				{
+					Entry:   zapcore.Entry{Level: DPanicLevel, Message: "foobar"},
+					Context: []Field{},
+				},
+			},
+			recoverValue: nil,
+		},
+		{
+			msg:      "dpanic no panic in production mode with panic hook",
+			level:    DPanicLevel,
+			opts:     opts(WithPanicHook(zapcore.WriteThenPanic)),
+			finished: true,
+			want: []observer.LoggedEntry{
+				{
+					Entry:   zapcore.Entry{Level: DPanicLevel, Message: "foobar"},
+					Context: []Field{},
+				},
+			},
+			recoverValue: nil,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.msg, func(t *testing.T) {
+			withLogger(t, InfoLevel, tt.opts, func(logger *Logger, logs *observer.ObservedLogs) {
+				var finished bool
+				recovered := make(chan any)
+				go func() {
+					defer func() {
+						recovered <- recover()
+					}()
+
+					logger.Log(tt.level, "foobar")
+					finished = true
+				}()
+
+				assert.Equal(t, tt.recoverValue, <-recovered, "unexpected value from recover()")
+				assert.Equal(t, tt.finished, finished, "expect goroutine finished state doesn't match")
+				assert.Equal(t, tt.want, logs.AllUntimed(), "unexpected logs")
+			})
+		})
+	}
 }
 
 func TestLoggerCustomOnFatal(t *testing.T) {
@@ -579,6 +1004,23 @@ func TestLoggerCustomOnFatal(t *testing.T) {
 	}
 }
 
+type customWriteHook struct {
+	called bool
+}
+
+func (h *customWriteHook) OnWrite(_ *zapcore.CheckedEntry, _ []Field) {
+	h.called = true
+}
+
+func TestLoggerWithFatalHook(t *testing.T) {
+	var h customWriteHook
+	withLogger(t, InfoLevel, opts(WithFatalHook(&h)), func(logger *Logger, logs *observer.ObservedLogs) {
+		logger.Fatal("great sadness")
+		assert.True(t, h.called)
+		assert.Equal(t, 1, logs.FilterLevelExact(FatalLevel).Len())
+	})
+}
+
 func TestNopLogger(t *testing.T) {
 	logger := NewNop()
 
@@ -597,6 +1039,16 @@ func TestNopLogger(t *testing.T) {
 		assert.Panics(t, func() {
 			logger.Panic("great sadness")
 		}, "Nop logger should still cause panics.")
+	})
+}
+
+func TestMust(t *testing.T) {
+	t.Run("must without an error does not panic", func(t *testing.T) {
+		assert.NotPanics(t, func() { Must(NewNop(), nil) }, "must paniced with no error")
+	})
+
+	t.Run("must with an error panics", func(t *testing.T) {
+		assert.Panics(t, func() { Must(nil, errors.New("an error")) }, "must did not panic with an error")
 	})
 }
 
