@@ -23,7 +23,6 @@ package zap_test
 import (
 	"bytes"
 	"encoding/json"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -98,7 +97,7 @@ func TestStacktraceFiltersVendorZap(t *testing.T) {
 
 		testDir := filepath.Join(goPath, "src/go.uber.org/zap_test/")
 		vendorDir := filepath.Join(testDir, "vendor")
-		require.NoError(t, os.MkdirAll(testDir, 0777), "Failed to create source director")
+		require.NoError(t, os.MkdirAll(testDir, 0o777), "Failed to create source director")
 
 		curFile := getSelfFilename(t)
 		setupSymlink(t, curFile, filepath.Join(testDir, curFile))
@@ -161,12 +160,8 @@ func verifyNoZap(t *testing.T, logs string) {
 }
 
 func withGoPath(t *testing.T, f func(goPath string)) {
-	goPath, err := ioutil.TempDir("", "gopath")
-	require.NoError(t, err, "Failed to create temporary directory for GOPATH")
-	//defer os.RemoveAll(goPath)
-
-	os.Setenv("GOPATH", goPath)
-	defer os.Setenv("GOPATH", os.Getenv("GOPATH"))
+	goPath := filepath.Join(t.TempDir(), "gopath")
+	t.Setenv("GOPATH", goPath)
 
 	f(goPath)
 }
@@ -180,7 +175,7 @@ func getSelfFilename(t *testing.T) string {
 
 func setupSymlink(t *testing.T, src, dst string) {
 	// Make sure the destination directory exists.
-	os.MkdirAll(filepath.Dir(dst), 0777)
+	require.NoError(t, os.MkdirAll(filepath.Dir(dst), 0o777))
 
 	// Get absolute path of the source for the symlink, otherwise we can create a symlink
 	// that uses relative paths.
