@@ -25,6 +25,5 @@ package tools
 
 import (
 	// Tools we use during development.
-	_ "golang.org/x/lint/golint"
-	_ "honnef.co/go/tools/cmd/staticcheck"
+	_ "golang.org/x/vuln/cmd/govulncheck"
 )
