@@ -29,10 +29,13 @@ import (
 	"testing"
 	"testing/quick"
 	"time"
+	"unicode/utf8"
 
+	"go.uber.org/zap/buffer"
 	"go.uber.org/zap/internal/bufferpool"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"go.uber.org/multierr"
 )
 
@@ -249,7 +252,7 @@ func TestJSONEncoderObjectFields(t *testing.T) {
 			desc:     "object (no nested namespace)",
 			expected: `"obj":{"obj-out":"obj-outside-namespace"},"not-obj":"should-be-outside-obj"`,
 			f: func(e Encoder) {
-				e.AddObject("obj", maybeNamespace{false})
+				assert.NoError(t, e.AddObject("obj", maybeNamespace{false}))
 				e.AddString("not-obj", "should-be-outside-obj")
 			},
 		},
@@ -257,7 +260,7 @@ func TestJSONEncoderObjectFields(t *testing.T) {
 			desc:     "object (with nested namespace)",
 			expected: `"obj":{"obj-out":"obj-outside-namespace","obj-namespace":{"obj-in":"obj-inside-namespace"}},"not-obj":"should-be-outside-obj"`,
 			f: func(e Encoder) {
-				e.AddObject("obj", maybeNamespace{true})
+				assert.NoError(t, e.AddObject("obj", maybeNamespace{true}))
 				e.AddString("not-obj", "should-be-outside-obj")
 			},
 		},
@@ -265,7 +268,7 @@ func TestJSONEncoderObjectFields(t *testing.T) {
 			desc:     "multiple open namespaces",
 			expected: `"k":{"foo":1,"middle":{"foo":2,"inner":{"foo":3}}}`,
 			f: func(e Encoder) {
-				e.AddObject("k", ObjectMarshalerFunc(func(enc ObjectEncoder) error {
+				err := e.AddObject("k", ObjectMarshalerFunc(func(enc ObjectEncoder) error {
 					e.AddInt("foo", 1)
 					e.OpenNamespace("middle")
 					e.AddInt("foo", 2)
@@ -273,6 +276,7 @@ func TestJSONEncoderObjectFields(t *testing.T) {
 					e.AddInt("foo", 3)
 					return nil
 				}))
+				assert.NoError(t, err)
 			},
 		},
 	}
@@ -289,10 +293,11 @@ func TestJSONEncoderTimeFormats(t *testing.T) {
 
 	f := func(e Encoder) {
 		e.AddTime("k", date)
-		e.AddArray("a", ArrayMarshalerFunc(func(enc ArrayEncoder) error {
+		err := e.AddArray("a", ArrayMarshalerFunc(func(enc ArrayEncoder) error {
 			enc.AppendTime(date)
 			return nil
 		}))
+		assert.NoError(t, err)
 	}
 	tests := []struct {
 		desc     string
@@ -420,7 +425,7 @@ func TestJSONEncoderArrays(t *testing.T) {
 			desc:     "object (no nested namespace) then string",
 			expected: `[{"obj-out":"obj-outside-namespace"},"should-be-outside-obj",{"obj-out":"obj-outside-namespace"},"should-be-outside-obj"]`,
 			f: func(arr ArrayEncoder) {
-				arr.AppendObject(maybeNamespace{false})
+				assert.NoError(t, arr.AppendObject(maybeNamespace{false}))
 				arr.AppendString("should-be-outside-obj")
 			},
 		},
@@ -428,7 +433,7 @@ func TestJSONEncoderArrays(t *testing.T) {
 			desc:     "object (with nested namespace) then string",
 			expected: `[{"obj-out":"obj-outside-namespace","obj-namespace":{"obj-in":"obj-inside-namespace"}},"should-be-outside-obj",{"obj-out":"obj-outside-namespace","obj-namespace":{"obj-in":"obj-inside-namespace"}},"should-be-outside-obj"]`,
 			f: func(arr ArrayEncoder) {
-				arr.AppendObject(maybeNamespace{true})
+				assert.NoError(t, arr.AppendObject(maybeNamespace{true}))
 				arr.AppendString("should-be-outside-obj")
 			},
 		},
@@ -530,10 +535,13 @@ type turducken struct{}
 func (t turducken) MarshalLogObject(enc ObjectEncoder) error {
 	return enc.AddArray("ducks", ArrayMarshalerFunc(func(arr ArrayEncoder) error {
 		for i := 0; i < 2; i++ {
-			arr.AppendObject(ObjectMarshalerFunc(func(inner ObjectEncoder) error {
+			err := arr.AppendObject(ObjectMarshalerFunc(func(inner ObjectEncoder) error {
 				inner.AddString("in", "chicken")
 				return nil
 			}))
+			if err != nil {
+				return err
+			}
 		}
 		return nil
 	}))
@@ -656,4 +664,73 @@ func TestJSONQuick(t *testing.T) {
 	// Focus on ASCII strings.
 	check(asciiRoundTripsCorrectlyString)
 	check(asciiRoundTripsCorrectlyByteString)
+}
+
+var _stringLikeCorpus = []string{
+	"",
+	"foo",
+	"bar",
+	"a\nb",
+	"a\tb",
+	"a\\b",
+	`a"b`,
+}
+
+func FuzzSafeAppendStringLike_bytes(f *testing.F) {
+	for _, s := range _stringLikeCorpus {
+		f.Add([]byte(s))
+	}
+	f.Fuzz(func(t *testing.T, b []byte) {
+		if !utf8.Valid(b) {
+			t.Skip()
+		}
+
+		fuzzSafeAppendStringLike(t, string(b), func(buf *buffer.Buffer) {
+			safeAppendStringLike(
+				(*buffer.Buffer).AppendBytes,
+				utf8.DecodeRune,
+				buf,
+				b,
+			)
+		})
+	})
+}
+
+func FuzzSafeAppendStringLike_string(f *testing.F) {
+	for _, s := range _stringLikeCorpus {
+		f.Add(s)
+	}
+	f.Fuzz(func(t *testing.T, s string) {
+		if !utf8.ValidString(s) {
+			t.Skip()
+		}
+
+		fuzzSafeAppendStringLike(t, s, func(buf *buffer.Buffer) {
+			safeAppendStringLike(
+				(*buffer.Buffer).AppendString,
+				utf8.DecodeRuneInString,
+				buf,
+				s,
+			)
+		})
+	})
+}
+
+func fuzzSafeAppendStringLike(
+	t *testing.T,
+	want string,
+	writeString func(*buffer.Buffer),
+) {
+	t.Helper()
+
+	buf := bufferpool.Get()
+	defer buf.Free()
+
+	buf.AppendByte('"')
+	writeString(buf)
+	buf.AppendByte('"')
+
+	var got string
+	require.NoError(t, json.Unmarshal(buf.Bytes(), &got))
+	assert.Equal(t, want, got)
 }
