@@ -18,6 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+// Package zapio provides tools for interacting with IO streams through Zap.
 package zapio
 
 import (
@@ -36,15 +37,15 @@ import (
 // and you want to log the output using your existing logger configuration. For
 // example,
 //
-//   writer := &zapio.Writer{Log: logger, Level: zap.DebugLevel}
-//   defer writer.Close()
+//	writer := &zapio.Writer{Log: logger, Level: zap.DebugLevel}
+//	defer writer.Close()
 //
-//   cmd := exec.CommandContext(ctx, ...)
-//   cmd.Stdout = writer
-//   cmd.Stderr = writer
-//   if err := cmd.Run(); err != nil {
-//       return err
-//   }
+//	cmd := exec.CommandContext(ctx, ...)
+//	cmd.Stdout = writer
+//	cmd.Stderr = writer
+//	if err := cmd.Run(); err != nil {
+//	    return err
+//	}
 //
 // Writer must be closed when finished to flush buffered data to the logger.
 type Writer struct {
